refactor(ui): split settings (library & storage and system) pages into horizontal nav layouts with child routes
